+++ src/test/auxiliary/lint_output_format.rs
@@ -10,19 +10,22 @@
 
 #![crate_name="lint_output_format"]
 #![crate_type = "lib"]
+#![feature(staged_api)]
 #![staged_api]
+#![unstable(feature = "test_feature")]
 
-#[deprecated]
-pub fn foo() -> uint {
+#[stable(feature = "test_feature", since = "1.0.0")]
+#[deprecated(since = "1.0.0")]
+pub fn foo() -> usize {
     20
 }
 
-#[unstable]
-pub fn bar() -> uint {
+#[unstable(feature = "test_feature")]
+pub fn bar() -> usize {
     40
 }
 
-#[unstable]
-pub fn baz() -> uint {
+#[unstable(feature = "test_feature")]
+pub fn baz() -> usize {
     30
 }
