+++ src/test/auxiliary/svh-a-change-type-ret.rs
@@ -24,9 +24,9 @@ pub trait V {}
 impl U for () {}
 impl V for () {}
 
-static A_CONSTANT : int = 2;
+static A_CONSTANT : isize = 2;
 
-pub fn foo<T:U>(_: int) -> i64 {
+pub fn foo<T:U>(_: isize) -> i64 {
     3
 }
