+++ src/test/debuginfo/gdb-pretty-struct-and-enums.rs
@@ -9,6 +9,7 @@
 // except according to those terms.
 
 // ignore-windows failing on win32 bot
+// ignore-freebsd: output doesn't match
 // ignore-tidy-linelength
 // ignore-lldb
 // ignore-android: FIXME(#10381)
@@ -75,12 +76,14 @@
 // gdb-command: print none_check2
 // gdb-check:$18 = None
 
+#![allow(dead_code, unused_variables)]
+
 use self::CStyleEnum::{CStyleEnumVar1, CStyleEnumVar2, CStyleEnumVar3};
 use self::MixedEnum::{MixedEnumCStyleVar, MixedEnumTupleVar, MixedEnumStructVar};
 use self::NestedEnum::{NestedVariant1, NestedVariant2};
 
 struct RegularStruct {
-    the_first_field: int,
+    the_first_field: isize,
     the_second_field: f64,
     the_third_field: bool,
     the_fourth_field: &'static str,
@@ -138,8 +141,8 @@ fn main() {
     let mixed_enum_tuple_var = MixedEnumTupleVar(106, 107, false);
     let mixed_enum_struct_var = MixedEnumStructVar { field1: 108.5, field2: 109 };
 
-    let some = Some(110u);
-    let none: Option<int> = None;
+    let some = Some(110_usize);
+    let none: Option<isize> = None;
     let some_fat = Some("abc");
     let none_fat: Option<&'static str> = None;
 
@@ -176,7 +179,7 @@ fn main() {
         }
     };
 
-    let none_check1: Option<(uint, Vec<uint>)> = None;
+    let none_check1: Option<(usize, Vec<usize>)> = None;
     let none_check2: Option<String> = None;
 
     zzz(); // #break
