+++ src/test/run-pass/asm-in-out-operand.rs
@@ -8,13 +8,14 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+
 #![feature(asm)]
 
 #[cfg(any(target_arch = "x86", target_arch = "x86_64"))]
 unsafe fn next_power_of_2(n: u32) -> u32 {
     let mut tmp = n;
     asm!("dec $0" : "+rm"(tmp) :: "cc");
-    let mut shift = 1u;
+    let mut shift = 1_usize;
     while shift <= 16 {
         asm!(
             "shr %cl, $2
@@ -34,15 +35,15 @@ pub fn main() {
         assert_eq!(2147483648, next_power_of_2(2147483647));
     }
 
-    let mut y: int = 5;
-    let x: int;
+    let mut y: isize = 5;
+    let x: isize;
     unsafe {
         // Treat the output as initialization.
         asm!(
             "shl $2, $1
             add $3, $1
             mov $1, $0"
-            : "=r"(x), "+r"(y) : "i"(3u), "ir"(7u) : "cc"
+            : "=r"(x), "+r"(y) : "i"(3_usize), "ir"(7_usize) : "cc"
         );
     }
     assert_eq!(x, 47);
