+++ src/test/run-pass/associated-types-issue-20371.rs
@@ -11,6 +11,8 @@
 // Test that we are able to have an impl that defines an associated type
 // before the actual trait.
 
-impl X for f64 { type Y = int; }
-trait X {type Y; }
+// pretty-expanded FIXME #23616
+
+impl X for f64 { type Y = isize; }
+trait X { type Y; }
 fn main() {}
