+++ src/test/run-pass/binops.rs
@@ -44,9 +44,9 @@ fn test_bool() {
 
 fn test_ptr() {
     unsafe {
-        let p1: *const u8 = ::std::mem::transmute(0u);
-        let p2: *const u8 = ::std::mem::transmute(0u);
-        let p3: *const u8 = ::std::mem::transmute(1u);
+        let p1: *const u8 = ::std::mem::transmute(0_usize);
+        let p2: *const u8 = ::std::mem::transmute(0_usize);
+        let p3: *const u8 = ::std::mem::transmute(1_usize);
 
         assert_eq!(p1, p2);
         assert!(p1 != p3);
@@ -59,13 +59,13 @@ fn test_ptr() {
     }
 }
 
-#[derive(PartialEq, Show)]
+#[derive(PartialEq, Debug)]
 struct p {
-  x: int,
-  y: int,
+  x: isize,
+  y: isize,
 }
 
-fn p(x: int, y: int) -> p {
+fn p(x: isize, y: isize) -> p {
     p {
         x: x,
         y: y
@@ -78,8 +78,8 @@ fn test_class() {
 
   unsafe {
   println!("q = {:x}, r = {:x}",
-         (::std::mem::transmute::<*const p, uint>(&q)),
-         (::std::mem::transmute::<*const p, uint>(&r)));
+         (::std::mem::transmute::<*const p, usize>(&q)),
+         (::std::mem::transmute::<*const p, usize>(&r)));
   }
   assert_eq!(q, r);
   r.y = 17;
