+++ src/test/run-pass/const-struct.rs
@@ -10,8 +10,8 @@
 
 use std::cmp;
 
-#[derive(Show)]
-struct foo { a: int, b: int, c: int }
+#[derive(Debug)]
+struct foo { a: isize, b: isize, c: isize }
 
 impl cmp::PartialEq for foo {
     fn eq(&self, other: &foo) -> bool {
