+++ src/test/run-pass/double-ref.rs
@@ -8,34 +8,36 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+// pretty-expanded FIXME #23616
+
 fn check_expr() {
-    let _:         & uint =     &1u;
-    let _:       & & uint =    &&1u;
-    let _:     & & & uint =   &&&1u;
-    let _:     & & & uint =  & &&1u;
-    let _:   & & & & uint =  &&&&1u;
-    let _:   & & & & uint = & &&&1u;
-    let _: & & & & & uint = &&&&&1u;
+    let _:         & usize =     &1;
+    let _:       & & usize =    &&1;
+    let _:     & & & usize =   &&&1;
+    let _:     & & & usize =  & &&1;
+    let _:   & & & & usize =  &&&&1;
+    let _:   & & & & usize = & &&&1;
+    let _: & & & & & usize = &&&&&1;
 }
 
 fn check_ty() {
-    let _:     &uint =         & 1u;
-    let _:    &&uint =       & & 1u;
-    let _:   &&&uint =     & & & 1u;
-    let _:  & &&uint =     & & & 1u;
-    let _:  &&&&uint =   & & & & 1u;
-    let _: & &&&uint =   & & & & 1u;
-    let _: &&&&&uint = & & & & & 1u;
+    let _:     &usize =         & 1;
+    let _:    &&usize =       & & 1;
+    let _:   &&&usize =     & & & 1;
+    let _:  & &&usize =     & & & 1;
+    let _:  &&&&usize =   & & & & 1;
+    let _: & &&&usize =   & & & & 1;
+    let _: &&&&&usize = & & & & & 1;
 }
 
 fn check_pat() {
-    let     &_ =         & 1u;
-    let    &&_ =       & & 1u;
-    let   &&&_ =     & & & 1u;
-    let  & &&_ =     & & & 1u;
-    let  &&&&_ =   & & & & 1u;
-    let & &&&_ =   & & & & 1u;
-    let &&&&&_ = & & & & & 1u;
+    let     &_ =         & 1_usize;
+    let    &&_ =       & & 1_usize;
+    let   &&&_ =     & & & 1_usize;
+    let  & &&_ =     & & & 1_usize;
+    let  &&&&_ =   & & & & 1_usize;
+    let & &&&_ =   & & & & 1_usize;
+    let &&&&&_ = & & & & & 1_usize;
 }
 
 pub fn main() {}
