+++ src/test/run-pass/issue-3794.rs
@@ -15,9 +15,9 @@ trait T {
     fn print(&self);
 }
 
-#[derive(Show)]
+#[derive(Debug)]
 struct S {
-    s: int,
+    s: isize,
 }
 
 impl T for S {
