+++ src/test/run-pass/method-attributes.rs
@@ -9,7 +9,10 @@
 // except according to those terms.
 
 // pp-exact - Make sure we print all the attributes
+// pretty-expanded FIXME #23616
+
 #![allow(unused_attribute)]
+#![feature(custom_attribute)]
 
 #[frobable]
 trait frobable {
@@ -20,7 +23,7 @@ trait frobable {
 }
 
 #[int_frobable]
-impl frobable for int {
+impl frobable for isize {
     #[frob_attr1]
     fn frob(&self) {
         #![frob_attr2]
